Reparacion de listas

+++ app/src/main/java/com/fsd/proyectoedson10/MainActivity.kt
@@ -151,7 +151,7 @@ class MainActivity : AppCompatActivity() {
                         db.ListDAO().selectList(
                             db.ListDAO().selectByName(
                                 x.listName
-                            ).idList.toInt()
+                            ).idList
                         ).listColor.toInt()
                     )
                     true
